support blob

+++ src/include/globals.h
@@ -21,7 +21,7 @@ union RecordIDType {
 constexpr RecordIDType INVALID_RECORD_ID { std::numeric_limits<uint64_t>::max() };
 constexpr PageIDType INVALID_PAGE_ID { std::numeric_limits<uint32_t>::max() };
 
-constexpr uint32_t PAGE_SIZE { 4096 };
+constexpr size_t PAGE_SIZE { 4096 };
 
 constexpr uint16_t SERVER_PORT { 8888 };
